refactor(bin): OutcomeRow contract + main.rs split + aged compat purged
- --json outcome envelope: OutcomeRow { id, outcome: bare token,
  message?: string } with the full SourceStatus treatment (schemars ->
  committed schema -> generated Raycast TS; gen-contract regenerates
  both goldens). The folded "failed: <msg>" string is GONE — the
  original back-compat argument assumed store-installed extensions, but
  the extension is unpublished: the compat target was an empty set, so
  the clean shape ships in the free window. Once the extension hits the
  store, wire changes need a version handshake (sharp-edged in both
  CLAUDE.mds). manage-sources.tsx drops the prefix-test + regex for
  typed row.outcome / row.message.
- main.rs 871 -> ~245 lines, move-only: crash.rs (panic hook + issue
  URLs), logging.rs (subscriber bootstrap + rotation), sources_cli.rs
  (presenters + emitters, next to the model they present); codecov
  ignore list follows the moved files (denominator stable).
- Aged back-compat purged (maintainer rulings, populations too old):
  the _ascii_agents legacy-sentinel strip (v0.3.0 tap installs; a
  leftover entry is inert — 0.13.0 retires the docs surface too, #446);
  resolve_connected's migrate-default (v0.4-0.7 configs: absent
  [sources] now plainly = disconnected; the upgrader replays onboarding
  — which IS the connect flow; has_hooks demoted pub(crate)); the
  cli_json golden re-pinned to the new semantics; the Codex uninstall
  legacy-basename fallback (sentinel + installer shipped in the same
  first release — no released version ever wrote sentinel-less entries;
  a hand-written shim entry now survives uninstall, pinned).

Design-debt findings #7, #16 + back-compat rulings #1, #2 + the
empty-set adjudications.
