+++ crates/pixtuoid/src/setup.rs
@@ -1,13 +1,12 @@
 //! First-run detection for the cinematic onboarding (PR2).
 //!
 //! `pixtuoid run` (the TUI) plays a one-time "move-in" overlay on its first ever
-//! launch; the headless `pixtuoid setup [--yes]` presenter (in `main.rs`) is the
-//! scriptable twin (Raycast / CI / scripting). The signal lives here as ONE pure
+//! launch; the headless `pixtuoid setup [--yes]` presenter (in the binary's
+//! `sources_cli.rs`) is the scriptable twin (Raycast / CI / scripting). The signal lives here as ONE pure
 //! predicate so both surfaces — and their tests — agree on what "first run" means.
 //!
 //! `pub` (not `pub(crate)`) because the binary's `main.rs` is a separate crate
-//! from this lib and computes it in `build_run_config` (the same reason
-//! `install::has_hooks` is `pub`).
+//! from this lib and computes it in `build_run_config`.
 
 use std::path::Path;
 
@@ -24,12 +23,15 @@ use crate::config::AppConfig;
 /// succeed. The caller passes `!load_warnings.is_empty()` right after `load`
 /// (a missing file returns defaults WITHOUT a warning, so it stays a first run).
 ///
-/// The second arm mirrors `config::resolve_connected`'s MIGRATE condition (an
-/// absent `[sources]` table falls back to install-state defaults): a user who
-/// has never bound a source through the panel/CLI is, by the same token, a user
-/// who has never been onboarded. Once any connect/disconnect persists a flag the
-/// table is non-empty and onboarding never re-triggers — even a *disconnected*
-/// flag (`false`) counts as "seen", since the table is non-empty.
+/// The second arm matches `config::resolve_connected`'s plain default (since
+/// 0.12.0 an absent/empty `[sources]` table means NOTHING connected — the old
+/// v0.4–0.7 install-state migrate inference is gone): a user who has never
+/// bound a source through the panel/CLI is, by the same token, a user who has
+/// never been onboarded — including a v0.4–0.7 upgrader whose config predates
+/// the flags; onboarding replays for them and IS their re-connect flow. Once
+/// any connect/disconnect persists a flag the table is non-empty and
+/// onboarding never re-triggers — even a *disconnected* flag (`false`) counts
+/// as "seen", since the table is non-empty.
 pub fn is_first_run(cfg: &AppConfig, path: &Path, load_degraded: bool) -> bool {
     !load_degraded && (!path.exists() || cfg.sources.is_empty())
 }
@@ -61,11 +63,34 @@ mod tests {
     #[test]
     fn existing_config_with_no_sources_table_is_first_run() {
         // A config exists (a theme, say) but no `[sources]` flag was ever
-        // written — the migrate condition, so still a first run.
+        // written — never onboarded, so still a first run.
         let dir = tempfile::tempdir().unwrap();
         let path = dir.path().join("config.toml");
         std::fs::write(&path, "theme = \"dracula\"\n").unwrap();
         assert!(is_first_run(&cfg_with(&[]), &path, false));
+    }
+
+    // The 0.12.0 consequence of dropping resolve_connected's migrate
+    // inference: a v0.4–0.7 upgrader's config (exists, parses fine, no
+    // [sources] table) resolves to NOTHING connected AND reads as a first
+    // run — the onboarding wizard replays and they re-connect there
+    // (acceptable: onboarding IS the connect flow).
+    #[test]
+    fn v04_v07_upgrader_config_without_sources_replays_onboarding() {
+        let dir = tempfile::tempdir().unwrap();
+        let path = dir.path().join("config.toml");
+        std::fs::write(&path, "theme = \"cyberpunk\"\nmax-desks = 8\n").unwrap();
+        let mut warnings = Vec::new();
+        let cfg = crate::config::load(&path, &mut warnings);
+        assert!(warnings.is_empty(), "a healthy old config is not degraded");
+        assert!(
+            crate::config::resolve_connected(&cfg).is_empty(),
+            "no [sources] ⇒ nothing connected (plain default, no install-state inference)"
+        );
+        assert!(
+            is_first_run(&cfg, &path, !warnings.is_empty()),
+            "…and the empty table replays onboarding, the re-connect path"
+        );
     }
 
     #[test]
