+++ crates/pixtuoid/src/version.rs
@@ -58,6 +58,11 @@ pub fn release_notes(version: &str) -> Option<&'static [&'static str]> {
         // anchoring on a marker is whitespace-independent — matching the `match`
         // brace would silently break if the indentation ever shifted.
         // [bump-inject-here]
+        "0.12.0" => Some(&[
+            "A whole-codebase cleanup: 39 review findings fixed — sturdier session tracking (a delegating agent no longer hides its own pending permission; live sessions survive log-content lookalikes), fresher sprites after a project rename, and pathfinding that recovers when a blocked route reopens",
+            "The office is more honest under pressure — a full pantry re-steams your coffee, narrow floating windows seat exactly as many agents as fit, and the web hero hires only when a desk is truly free",
+            "Under the hood: the render engine's public API slimmed and semver-gated, per-CLI transcript knowledge fully routed through the source registry, and the daily security-audit signal revived",
+        ]),
         "0.11.1" => Some(&[
             "Maintenance release — the animated office is unchanged from 0.11.0; documentation polish across the site and READMEs, plus a supply-chain hardening of how pixtuoid ships",
             "Releases now publish to crates.io and npm via OIDC trusted publishing — CI holds no long-lived registry tokens, shrinking the supply-chain attack surface",
