+++ crates/pixtuoid/tests/wire_to_pixels.rs
@@ -484,8 +484,8 @@ fn lobster_px(buf: &pixtuoid_core::sprite::RgbBuffer) -> usize {
         }, // shade
     ];
     let mut n = 0;
-    for y in 0..buf.height {
-        for x in 0..buf.width {
+    for y in 0..buf.height() {
+        for x in 0..buf.width() {
             if reds.contains(&buf.get(x, y)) {
                 n += 1;
             }
