+++ integrations/raycast/.prettierignore
@@ -1,3 +1,4 @@
 # generated / contract artifacts — never reformat
 src/contract.ts
 contract/
+src/contract-outcome.ts
