+++ integrations/raycast/eslint.config.mjs
@@ -2,9 +2,10 @@ import { defineConfig } from "eslint/config";
 import raycastConfig from "@raycast/eslint-config";
 
 // Generated files — never lint them: `raycast-env.d.ts` + `dist/` come from
-// `ray build`; `src/contract.ts` is generated by `npm run gen:contract` from the
-// committed JSON Schema (the --json contract). Keeps `eslint .` clean in CI.
+// `ray build`; `src/contract.ts` + `src/contract-outcome.ts` are generated by
+// `npm run gen:contract` from the committed JSON Schemas (the --json contract).
+// Keeps `eslint .` clean in CI.
 export default defineConfig([
-  { ignores: ["raycast-env.d.ts", "dist/**", "src/contract.ts"] },
+  { ignores: ["raycast-env.d.ts", "dist/**", "src/contract.ts", "src/contract-outcome.ts"] },
   ...raycastConfig,
 ]);
