+++ integrations/raycast/package.json
@@ -72,7 +72,7 @@
     "dev": "ray develop",
     "lint": "ray lint",
     "fix-lint": "ray lint --fix",
-    "gen:contract": "json2ts -i contract/source-status.schema.json -o src/contract.ts",
+    "gen:contract": "json2ts -i contract/source-status.schema.json -o src/contract.ts && json2ts -i contract/outcome-row.schema.json -o src/contract-outcome.ts",
     "publish": "npx @raycast/api@latest publish"
   }
 }
