+++ integrations/raycast/src/contract-outcome.ts
@@ -1,0 +1,40 @@
+/* eslint-disable */
+/**
+ * This file was automatically generated by json-schema-to-typescript.
+ * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
+ * and run json-schema-to-typescript to regenerate this file.
+ */
+
+/**
+ * One row of the `--json` batch envelope `pixtuoid connect|disconnect|sources set`
+ * print — the SECOND stable wire contract the Raycast extension parses (alongside
+ * `SourceStatus`), with the same treatment: a committed JSON Schema
+ * (`integrations/raycast/contract/outcome-row.schema.json`, golden-tested below)
+ * the extension's TS type is generated from (`gen:contract`). The wire shape is
+ * `{id, outcome, message?}` — a bare machine token plus an optional human-detail
+ * field, split from the older folded `failed: <msg>` string BEFORE the
+ * extension's store publication (the in-repo extension ships atomically with
+ * the binary, so no installed consumer parsed the old form). Once the extension
+ * is store-published, changing this shape needs a version handshake — see the
+ * sharp edge in `crates/pixtuoid/CLAUDE.md`. Pinned by
+ * `outcome_row_json_shape_is_the_raycast_contract` + the envelope test in
+ * `sources_cli.rs`.
+ */
+export interface OutcomeRow {
+  /**
+   * The registry source id the outcome applies to (e.g. `codex`).
+   */
+  id: string;
+  /**
+   * Human-readable detail for the row — present exactly when the outcome
+   * carries any (`failed`), and OMITTED (not `null`) otherwise, so a
+   * success row stays the minimal `{id, outcome}`.
+   */
+  message?: string | null;
+  /**
+   * The BARE outcome token from `ChangeOutcome::wire_token`: `connected` |
+   * `disconnected` | `no_op` | `failed`. Machine-matchable (compare with
+   * `===`), never carrying human text — the detail rides in `message`.
+   */
+  outcome: string;
+}
