+++ integrations/raycast/src/manage-sources.tsx
@@ -38,10 +38,10 @@ export default function ManageSources() {
     });
     try {
       const row = await toggleSource(item.id, item.connected);
-      if (row?.outcome?.startsWith("failed")) {
+      if (row?.outcome === "failed") {
         toast.style = Toast.Style.Failure;
         toast.title = `${item.display_name}: failed`;
-        toast.message = row.outcome.replace(/^failed:\s*/, "");
+        toast.message = row.message ?? undefined;
       } else {
         toast.style = Toast.Style.Success;
         toast.title = `${item.display_name} — ${row?.outcome ?? "updated"}`;
